subindo correção de atividades e implementações a mais

+++ src/main/java/com/tcc/e_language_api/entity/PerfilIdioma.java
@@ -39,5 +39,5 @@ public class PerfilIdioma {
             ))
     private NivelIdioma nivelIdioma;
     @Column(name = "pontos_ranking", nullable = false)
-    private int pontosRanking;
+    private Double pontosRanking;
 }
